apps/reader/pyproject.toml
[project]
name = "agenttalks-reader"
version = "0.1.0"
description = "Read and summarizes content for the agenttalks project"
readme = "README.md"
authors = [
    { name = "Willem Meints", email = "1550763+wmeints@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15.4",
    "pydantic>=2.10.5",
    "dapr-agents>=0.5.1",
    "fastapi>=0.115.12",
    "dapr-ext-fastapi>=1.15.0",
    "uvicorn>=0.34.2",
    "dapr-ext-workflow>=1.15.0",
]

[project.scripts]
agenttalks-reader = "agenttalks.reader.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agenttalks"]

[dependency-groups]
dev = [
    "callee>=0.3.1",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
]
